Revert "Update type based on BVH::LevelInside"
This reverts commit e89c24974f3b9b2a761374ae953b97210b6c5072.

+++ src/geocel/vg/VecgeomTypes.hh
@@ -63,7 +63,7 @@ namespace celeritas
 //---------------------------------------------------------------------------//
 
 using VgSurfaceInt = long;
-using VgPlacedVolumeInt = long;
+using VgPlacedVolumeInt = int;
 using vg_real_type = VECGEOM_PRECISION_NAMESPACE::Precision;
 
 #if defined(VECGEOM_BVH_SINGLE) || defined(__DOXYGEN__)
